Add files via upload
Neural Network SMS Text Classifier - Google Colaboratory
In this challenge, i created a machine learning model that classify SMS messages as either "ham" or "spam". A "ham" message is a normal message sent by a friend. A "spam" message is an advertisement or a message sent by a company.For this challenge, you i used the SMS Spam Collection dataset. The dataset has already been grouped into train data and test data. i created a function called predict_message that takes a message string as an argument and returns a list.